Move CLI progress bar constant into manual helpers directory
