+++ src/cli/manual/progress-bar-constants.js
