+++ src/cli/options/progress-bar.js
@@ -2,7 +2,7 @@ import {
     coercePositiveInteger,
     resolveIntegerOption
 } from "./integer-utils.js";
-import { DEFAULT_PROGRESS_BAR_WIDTH } from "./progress-bar-constants.js";
+import { DEFAULT_PROGRESS_BAR_WIDTH } from "../manual/progress-bar-constants.js";
 
 const createWidthErrorMessage = (received) =>
     `Progress bar width must be a positive integer (received ${received}).`;
